pyproject.toml
[project]
name = "multi-doc-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt==0.9",
    "dvc>=3.66.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.110",
    "fastembed>=0.7.4",
    "ipykernel==6.30.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain-openai>=0.1.0",
    "langsmith>=0.5.0",
    "pydantic>=2.5",
    "pypdf>=6.6.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "structlog>=24.0.0",
    "uvicorn>=0.27",
]
